requirements.txt
Django==3.2.9
gunicorn==20.1.0
pipenv==2022.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-telegram-bot==13.11
requests==2.27.1
virtualenv==20.14.1
transliterate==1.10.2